move ParquetJoinTest logic to ParquetRewriterTest
