+++ parquet-hadoop/src/test/java/org/apache/parquet/hadoop/rewrite/ParquetRewriterTest.java
@@ -36,15 +36,10 @@ import static org.junit.Assert.assertTrue;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
+import java.util.*;
+import java.util.function.BiFunction;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.Path;
 import org.apache.parquet.HadoopReadOptions;
@@ -108,6 +103,7 @@ public class ParquetRewriterTest {
   private final boolean usingHadoop;
 
   private List<EncryptionTestFile> inputFiles = null;
+  private List<List<EncryptionTestFile>> inputFilesR = null;
   private String outputFile = null;
   private ParquetRewriter rewriter = null;
 
@@ -175,6 +171,7 @@ public class ParquetRewriterTest {
   @Before
   public void setUp() {
     outputFile = TestFileBuilder.createTempFile("test");
+    inputFilesR = new ArrayList<>();
   }
 
   @Test
@@ -725,6 +722,90 @@ public class ParquetRewriterTest {
         .build());
   }
 
+  @Test
+  public void testStitchThreeInputsDifferentRowGroupSize() throws Exception {
+    testThreeInputsDifferentRowGroupSize();
+
+    // Only merge two files but do not change anything.
+    List<Path> inputPathsL = inputFiles.stream()
+        .map(x -> new Path(x.getFileName()))
+        .collect(Collectors.toList());
+    List<List<Path>> inputPathsR = inputFilesR.stream()
+        .map(x -> x.stream().map(y -> new Path(y.getFileName())).collect(Collectors.toList()))
+        .collect(Collectors.toList());
+    RewriteOptions.Builder builder = createBuilder(inputPathsL, inputPathsR);
+    RewriteOptions options = builder.indexCacheStrategy(indexCacheStrategy).build();
+
+    rewriter = new ParquetRewriter(options, true);
+    rewriter.processBlocks();
+    rewriter.close();
+
+    // Verify the schema are not changed
+    ParquetMetadata pmd =
+        ParquetFileReader.readFooter(conf, new Path(outputFile), ParquetMetadataConverter.NO_FILTER);
+    MessageType schema = pmd.getFileMetaData().getSchema();
+    MessageType expectSchema = createSchema();
+    assertEquals(expectSchema, schema);
+
+    // Verify the merged data are not changed
+    validateColumnData(Collections.emptySet(), Collections.emptySet(), null);
+  }
+
+  private void testThreeInputsDifferentRowGroupSize() throws IOException {
+    inputFiles = Lists.newArrayList(
+        new TestFileBuilder(conf, createSchemaL())
+            .withNumRecord(numRecord / 2)
+            .withRowGroupSize(5_000_000)
+            .withCodec("GZIP")
+            .withPageSize(ParquetProperties.DEFAULT_PAGE_SIZE)
+            .withWriterVersion(writerVersion)
+            .build(),
+        new TestFileBuilder(conf, createSchemaL())
+            .withNumRecord(numRecord - (numRecord / 2))
+            .withRowGroupSize(6_000_000)
+            .withCodec("GZIP")
+            .withPageSize(ParquetProperties.DEFAULT_PAGE_SIZE)
+            .withWriterVersion(writerVersion)
+            .build()
+    );
+    inputFilesR = Lists.newArrayList(
+        Lists.newArrayList(
+            Lists.newArrayList(
+                new TestFileBuilder(conf, createSchemaR1())
+                    .withNumRecord(numRecord)
+                    .withRowGroupSize(7_000_000)
+                    .withCodec("UNCOMPRESSED")
+                    .withPageSize(ParquetProperties.DEFAULT_PAGE_SIZE)
+                    .withWriterVersion(writerVersion)
+                    .build()
+            ),
+            Lists.newArrayList(
+                new TestFileBuilder(conf, createSchemaR2())
+                    .withNumRecord(numRecord / 3)
+                    .withRowGroupSize(200_000)
+                    .withCodec("UNCOMPRESSED")
+                    .withPageSize(ParquetProperties.DEFAULT_PAGE_SIZE)
+                    .withWriterVersion(writerVersion)
+                    .build(),
+                new TestFileBuilder(conf, createSchemaR2())
+                    .withNumRecord(numRecord / 3)
+                    .withRowGroupSize(300_000)
+                    .withCodec("UNCOMPRESSED")
+                    .withPageSize(ParquetProperties.DEFAULT_PAGE_SIZE)
+                    .withWriterVersion(writerVersion)
+                    .build(),
+                new TestFileBuilder(conf, createSchemaR2())
+                    .withNumRecord(numRecord - 2 * (numRecord / 3))
+                    .withRowGroupSize(400_000)
+                    .withCodec("UNCOMPRESSED")
+                    .withPageSize(ParquetProperties.DEFAULT_PAGE_SIZE)
+                    .withWriterVersion(writerVersion)
+                    .build()
+            )
+        )
+    );
+  }
+
   private MessageType createSchema() {
     return new MessageType(
         "schema",
@@ -738,6 +819,33 @@ public class ParquetRewriterTest {
             "Links",
             new PrimitiveType(REPEATED, BINARY, "Backward"),
             new PrimitiveType(REPEATED, BINARY, "Forward")));
+  }
+
+
+  private MessageType createSchemaL() {
+    return new MessageType(
+        "schema",
+        new PrimitiveType(OPTIONAL, INT64, "DocId"),
+        new PrimitiveType(REQUIRED, BINARY, "Name"),
+        new PrimitiveType(OPTIONAL, BINARY, "Gender"),
+        new PrimitiveType(REPEATED, FLOAT, "FloatFraction"),
+        new PrimitiveType(OPTIONAL, DOUBLE, "DoubleFraction"));
+  }
+
+  private MessageType createSchemaR1() {
+    return new MessageType(
+        "schema",
+        new GroupType(
+            OPTIONAL,
+            "Links",
+            new PrimitiveType(REPEATED, BINARY, "Backward"),
+            new PrimitiveType(REPEATED, BINARY, "Forward")));
+  }
+
+  private MessageType createSchemaR2() {
+    return new MessageType(
+        "schema",
+        new PrimitiveType(REPEATED, FLOAT, "FloatFraction"));
   }
 
   private void validateColumnData(
@@ -754,38 +862,51 @@ public class ParquetRewriterTest {
       totalRows += inputFile.getFileContent().length;
     }
 
+    List<List<SimpleGroup>> fileContents = Stream.concat(
+            Stream.of(inputFiles),
+            inputFilesR.stream()
+        ).map(x -> x.stream().flatMap(y -> Arrays.stream(y.getFileContent())).collect(Collectors.toList()))
+        .collect(Collectors.toList());
+    BiFunction<String, Integer, Group> groups = (name, rowIdx) -> {
+      for (int i = fileContents.size() - 1; i >= 0; i--) {
+        SimpleGroup expGroup = fileContents.get(i).get(rowIdx);
+        GroupType fileSchema = expGroup.getType();
+        if (fileSchema.containsField(name)) {
+          return expGroup;
+        }
+      }
+      throw new IllegalStateException("Group '"+name+"' at position "+rowIdx+" was not found in input files!");
+    };
     for (int i = 0; i < totalRows; i++) {
       Group group = reader.read();
       assertNotNull(group);
-
-      SimpleGroup expectGroup = inputFiles.get(i / numRecord).getFileContent()[i % numRecord];
 
       if (!prunePaths.contains("DocId")) {
         if (nullifiedPaths.contains("DocId")) {
           assertThrows(RuntimeException.class, () -> group.getLong("DocId", 0));
         } else {
-          assertEquals(group.getLong("DocId", 0), expectGroup.getLong("DocId", 0));
+          assertEquals(group.getLong("DocId", 0), groups.apply("DocId", i).getLong("DocId", 0));
         }
       }
 
       if (!prunePaths.contains("Name") && !nullifiedPaths.contains("Name")) {
         assertArrayEquals(
             group.getBinary("Name", 0).getBytes(),
-            expectGroup.getBinary("Name", 0).getBytes());
+            groups.apply("Name", i).getBinary("Name", 0).getBytes());
       }
 
       if (!prunePaths.contains("Gender") && !nullifiedPaths.contains("Gender")) {
         assertArrayEquals(
             group.getBinary("Gender", 0).getBytes(),
-            expectGroup.getBinary("Gender", 0).getBytes());
+            groups.apply("Gender", i).getBinary("Gender", 0).getBytes());
       }
 
       if (!prunePaths.contains("FloatFraction") && !nullifiedPaths.contains("FloatFraction")) {
-        assertEquals(group.getFloat("FloatFraction", 0), expectGroup.getFloat("FloatFraction", 0), 0);
+        assertEquals(group.getFloat("FloatFraction", 0), groups.apply("FloatFraction", i).getFloat("FloatFraction", 0), 0);
       }
 
       if (!prunePaths.contains("DoubleFraction") && !nullifiedPaths.contains("DoubleFraction")) {
-        assertEquals(group.getDouble("DoubleFraction", 0), expectGroup.getDouble("DoubleFraction", 0), 0);
+        assertEquals(group.getDouble("DoubleFraction", 0), groups.apply("DoubleFraction", i).getDouble("DoubleFraction", 0), 0);
       }
 
       Group subGroup = group.getGroup("Links", 0);
@@ -793,7 +914,7 @@ public class ParquetRewriterTest {
       if (!prunePaths.contains("Links.Backward") && !nullifiedPaths.contains("Links.Backward")) {
         assertArrayEquals(
             subGroup.getBinary("Backward", 0).getBytes(),
-            expectGroup
+            groups.apply("Links", i)
                 .getGroup("Links", 0)
                 .getBinary("Backward", 0)
                 .getBytes());
@@ -805,7 +926,7 @@ public class ParquetRewriterTest {
         } else {
           assertArrayEquals(
               subGroup.getBinary("Forward", 0).getBytes(),
-              expectGroup
+              groups.apply("Links", i)
                   .getGroup("Links", 0)
                   .getBinary("Forward", 0)
                   .getBytes());
@@ -1042,16 +1163,22 @@ public class ParquetRewriterTest {
   }
 
   private RewriteOptions.Builder createBuilder(List<Path> inputPaths) throws IOException {
+    return createBuilder(inputPaths, new ArrayList<>());
+  }
+
+  private RewriteOptions.Builder createBuilder(List<Path> inputPathsL, List<List<Path>> inputPathsR) throws IOException {
     RewriteOptions.Builder builder;
     if (usingHadoop) {
       Path outputPath = new Path(outputFile);
-      builder = new RewriteOptions.Builder(conf, inputPaths, outputPath);
+      builder = new RewriteOptions.Builder(conf, inputPathsL, outputPath);
+      inputPathsR.forEach(builder::addInputPathsR);
     } else {
       OutputFile outputPath = HadoopOutputFile.fromPath(new Path(outputFile), conf);
-      List<InputFile> inputs = inputPaths.stream()
+      List<InputFile> inputsL = inputPathsL.stream()
           .map(p -> HadoopInputFile.fromPathUnchecked(p, conf))
           .collect(Collectors.toList());
-      builder = new RewriteOptions.Builder(parquetConf, inputs, outputPath);
+      builder = new RewriteOptions.Builder(parquetConf, inputsL, outputPath);
+      inputPathsR.forEach(builder::addInputPathsR);
     }
     return builder;
   }
@@ -1072,4 +1199,5 @@ public class ParquetRewriterTest {
     assertEquals(subFields.get(0).getName(), "Backward");
     assertEquals(subFields.get(1).getName(), "Forward");
   }
+
 }
